Check attestation target checkpoint is within current or previous epoch
