+++ beacon-chain/blockchain/receive_attestation.go
@@ -13,7 +13,6 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/params"
 	"github.com/prysmaticlabs/prysm/shared/slotutil"
-	"github.com/prysmaticlabs/prysm/shared/timeutils"
 	"github.com/sirupsen/logrus"
 	"go.opencensus.io/trace"
 )
@@ -127,7 +126,7 @@ func (s *Service) processAttestation(subscribedToStateEvents chan struct{}) {
 					log.WithError(err).Error("Could not delete fork choice attestation in pool")
 				}
 
-				if !s.verifyCheckpointEpoch(a.Data.Target) {
+				if !helpers.VerifyCheckpointEpoch(a.Data.Target, s.genesisTime) {
 					continue
 				}
 
@@ -143,24 +142,4 @@ func (s *Service) processAttestation(subscribedToStateEvents chan struct{}) {
 			}
 		}
 	}
-}
-
-// This verifies the epoch of input checkpoint is within current epoch and previous epoch
-// with respect to current time. Returns true if it's within, false if it's not.
-func (s *Service) verifyCheckpointEpoch(c *ethpb.Checkpoint) bool {
-	now := uint64(timeutils.Now().Unix())
-	genesisTime := uint64(s.genesisTime.Unix())
-	currentSlot := (now - genesisTime) / params.BeaconConfig().SecondsPerSlot
-	currentEpoch := helpers.SlotToEpoch(currentSlot)
-
-	var prevEpoch uint64
-	if currentEpoch > 1 {
-		prevEpoch = currentEpoch - 1
-	}
-
-	if c.Epoch != prevEpoch && c.Epoch != currentEpoch {
-		return false
-	}
-
-	return true
 }
