+++ beacon-chain/sync/validate_aggregate_proof.go
@@ -44,14 +44,17 @@ func (s *Service) validateAggregateAndProof(ctx context.Context, pid peer.ID, ms
 	if !ok {
 		return pubsub.ValidationReject
 	}
+	if m.Message == nil || m.Message.Aggregate == nil || m.Message.Aggregate.Data == nil {
+		return pubsub.ValidationReject
+	}
 	if err := helpers.ValidateAttestationTime(m.Message.Aggregate.Data.Slot, s.chain.GenesisTime()); err != nil {
 		traceutil.AnnotateError(span, err)
 		return pubsub.ValidationIgnore
 	}
-
-	if m.Message == nil || m.Message.Aggregate == nil || m.Message.Aggregate.Data == nil {
-		return pubsub.ValidationReject
+	if !helpers.VerifyCheckpointEpoch(m.Message.Aggregate.Data.Target, s.chain.GenesisTime()) {
+		return pubsub.ValidationIgnore
 	}
+
 	// Verify this is the first aggregate received from the aggregator with index and slot.
 	if s.hasSeenAggregatorIndexEpoch(m.Message.Aggregate.Data.Target.Epoch, m.Message.AggregatorIndex) {
 		return pubsub.ValidationIgnore
